backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2>=2.9
python-dotenv
gunicorn